No tests completed; draft of 0.14.1
Note that Stata might not handle weights accurately in pctile and xtile
(or, at the very least, it does not seem to follow its documented
formulas in all cases). This is somewhat strange and since the
underlying code is in `_pctile`, which is built in, it is not possible
to examine it. The `collapse` code, which is available, looks correct
and gives the expected answers (which can differ from `_pctile` when
using weights).