Windows passing, updated plugin; re-sort of strL vars is not stable
When presented with sorted data, Stata's `sort` usually does not re-sort
it.  However, it seems to re-run its algorithm when sorting strL
data. This means that I need to use a stable sort for strL variables
because they would be re-run through the algorith. Hopefully this does
not affect performance too much.